One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/planning/rsassignment/rsassignment.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RsassignmentComponent } from './rsassignment.component';

describe('Component: Rsassignment', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RsassignmentComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RsassignmentComponent],
      (component: RsassignmentComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RsassignmentComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RsassignmentComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-rsassignment></app-rsassignment>
  `,
  directives: [RsassignmentComponent]
})
class RsassignmentComponentTestController {
}

